Smart contract for simple energy sharing concept
This is a demo application for smart contract in energy sharing. Each
node will be considered as a meter in local neighborhood. They will be
initialed a 100 tokens for trading energy. Meter’s will post request to
sell and buy.

The example is running under testrpc and truffle framework. The
instruction to run is in read.md